requirements.txt
dbt-core==1.9.8
dbt-snowflake==1.9.1
snowflake-connector-python>=3.0.0
